Implement list of new updates
Add load and save functions to read/write from/to a file in the
directory(passengers.dat). Update main to use these functions. Create an
enumerated type for payment preference as well as update existing
functions to read and write from it. Change incrementation of Passenger
ID. Create a decrement function for future delete function.