Fix multiple output bugs and remove test output
Fixed a bug where a driver can still be reserved for rides with an
active ride. Solve an issue where if a driver had a cancelled ride with
the same id only the already cancelled ride would be set to cancelled.
Could be a data format error as opening the data files found some
inaccurate artifacts from potentially editing the save and load
functions.